Fix type mixup for ulong/ushort/short.

+++ TextTableBuilder/TypeHandlers/ULongHandler.cs
@@ -2,5 +2,5 @@
 
 public class ULongHandler : NaturalNumberHandler
 {
-    public override string Handle(object value, IFormatProvider formatProvider) => ((short)value).ToString(Format, formatProvider);
+    public override string Handle(object value, IFormatProvider formatProvider) => ((ulong)value).ToString(Format, formatProvider);
 }
